reorganize source code to be easier to navigate
