+++ cryojax/experimental/__init__.py
@@ -1,14 +1,12 @@
-from ..simulator._multislice_integrator import (
+from ..simulator._multislice import (
     AbstractMultisliceIntegrator as AbstractMultisliceIntegrator,
     FFTMultisliceIntegrator as FFTMultisliceIntegrator,
-)
-from ..simulator._scattering_theory import (
     MultisliceScatteringTheory as MultisliceScatteringTheory,
 )
 from ..simulator._solvent_2d import (
     GRFSolvent2D as GRFSolvent2D,
     SolventMixturePower as SolventMixturePower,
 )
-from ..simulator._volume_integrator import (
+from ..simulator._volume import (
     EwaldSphereExtraction as EwaldSphereExtraction,
 )
