--- cryojax/simulator/_image_model/__init__.py
@@ -1,11 +1,0 @@
-from .base_image_model import (
-    AbstractImageModel as AbstractImageModel,
-    LinearImageModel as LinearImageModel,
-    ProjectionImageModel as ProjectionImageModel,
-)
-from .physical_image_model import (
-    AbstractPhysicalImageModel as AbstractPhysicalImageModel,
-    ContrastImageModel as ContrastImageModel,
-    ElectronCountsImageModel as ElectronCountsImageModel,
-    IntensityImageModel as IntensityImageModel,
-)
