+++ cryojax/simulator/_multislice/__init__.py
@@ -1,0 +1,7 @@
+from .multislice_integrator import (
+    AbstractMultisliceIntegrator as AbstractMultisliceIntegrator,
+    FFTMultisliceIntegrator as FFTMultisliceIntegrator,
+)
+from .multislice_scattering_theory import (
+    MultisliceScatteringTheory as MultisliceScatteringTheory,
+)
