--- cryojax/simulator/_multislice_integrator/__init__.py
@@ -1,4 +1,0 @@
-from .base_multislice_integrator import (
-    AbstractMultisliceIntegrator as AbstractMultisliceIntegrator,
-)
-from .fft_multislice_integrator import FFTMultisliceIntegrator as FFTMultisliceIntegrator
