--- cryojax/simulator/_scattering_theory/__init__.py
@@ -1,13 +1,0 @@
-from .base_scattering_theory import (
-    AbstractScatteringTheory as AbstractScatteringTheory,
-    AbstractWaveScatteringTheory as AbstractWaveScatteringTheory,
-)
-from .multislice_scattering_theory import (
-    MultisliceScatteringTheory as MultisliceScatteringTheory,
-)
-from .strong_phase_scattering_theory import (
-    StrongPhaseScatteringTheory as StrongPhaseScatteringTheory,
-)
-from .weak_phase_scattering_theory import (
-    WeakPhaseScatteringTheory as WeakPhaseScatteringTheory,
-)
