+++ cryojax/simulator/_volume/__init__.py
@@ -1,13 +1,24 @@
-from .base_volume import (
-    AbstractVolumeParametrization as AbstractVolumeParametrization,
-    AbstractVolumeRepresentation as AbstractVolumeRepresentation,
-)
-from .representations import (
-    AbstractAtomVolume as AbstractAtomVolume,
-    AbstractVoxelVolume as AbstractVoxelVolume,
-    FourierVoxelGridVolume as FourierVoxelGridVolume,
-    FourierVoxelSplineVolume as FourierVoxelSplineVolume,
+from .atoms import (
+    FFTAtomProjection as FFTAtomProjection,
+    FFTAtomRenderFn as FFTAtomRenderFn,
+    GaussianMixtureProjection as GaussianMixtureProjection,
+    GaussianMixtureRenderFn as GaussianMixtureRenderFn,
     GaussianMixtureVolume as GaussianMixtureVolume,
     IndependentAtomVolume as IndependentAtomVolume,
+)
+from .base_volume import (
+    AbstractAtomVolume as AbstractAtomVolume,
+    AbstractVolumeIntegrator as AbstractVolumeIntegrator,
+    AbstractVolumeParametrization as AbstractVolumeParametrization,
+    AbstractVolumeRenderFn as AbstractVolumeRenderFn,
+    AbstractVolumeRepresentation as AbstractVolumeRepresentation,
+    AbstractVoxelVolume as AbstractVoxelVolume,
+)
+from .voxels import (
+    EwaldSphereExtraction as EwaldSphereExtraction,
+    FourierSliceExtraction as FourierSliceExtraction,
+    FourierVoxelGridVolume as FourierVoxelGridVolume,
+    FourierVoxelSplineVolume as FourierVoxelSplineVolume,
     RealVoxelGridVolume as RealVoxelGridVolume,
+    RealVoxelProjection as RealVoxelProjection,
 )
