+++ cryojax/simulator/_volume/atoms/__init__.py
@@ -1,0 +1,10 @@
+from .gaussian_volume import (
+    GaussianMixtureProjection as GaussianMixtureProjection,
+    GaussianMixtureRenderFn as GaussianMixtureRenderFn,
+    GaussianMixtureVolume as GaussianMixtureVolume,
+)
+from .independent_atom_volume import (
+    FFTAtomProjection as FFTAtomProjection,
+    FFTAtomRenderFn as FFTAtomRenderFn,
+    IndependentAtomVolume as IndependentAtomVolume,
+)
