--- cryojax/simulator/_volume/representations/__init__.py
@@ -1,14 +1,0 @@
-from .base_representations import (
-    AbstractAtomVolume as AbstractAtomVolume,
-    AbstractVoxelVolume as AbstractVoxelVolume,
-)
-from .gaussian_volume import GaussianMixtureVolume as GaussianMixtureVolume
-from .independent_atom_volume import (
-    IndependentAtomVolume as IndependentAtomVolume,
-    PengScatteringFactor as PengScatteringFactor,
-)
-from .voxel_volume import (
-    FourierVoxelGridVolume as FourierVoxelGridVolume,
-    FourierVoxelSplineVolume as FourierVoxelSplineVolume,
-    RealVoxelGridVolume as RealVoxelGridVolume,
-)
