+++ cryojax/simulator/_volume/voxels/__init__.py
@@ -1,0 +1,10 @@
+from .fourier_voxels import (
+    EwaldSphereExtraction as EwaldSphereExtraction,
+    FourierSliceExtraction as FourierSliceExtraction,
+    FourierVoxelGridVolume as FourierVoxelGridVolume,
+    FourierVoxelSplineVolume as FourierVoxelSplineVolume,
+)
+from .real_voxels import (
+    RealVoxelGridVolume as RealVoxelGridVolume,
+    RealVoxelProjection as RealVoxelProjection,
+)
