+++ cryojax/simulator/_volume/voxels/real_voxels.py
@@ -1,18 +1,20 @@
 """
-Using non-uniform FFTs for computing volume projections.
+Real voxel-based representations of a volume.
 """
 
 import math
-from typing import Any, ClassVar
-from typing_extensions import override
+from typing import Any, ClassVar, cast
+from typing_extensions import Self, override
 
+import equinox as eqx
 import jax.numpy as jnp
 from jaxtyping import Array, Complex, Float
 
-from ...ndimage import convert_fftn_to_rfftn, irfftn
-from .._image_config import AbstractImageConfig
-from .._volume import RealVoxelGridVolume
-from .base_integrator import AbstractVolumeIntegrator
+from ....jax_util import NDArrayLike
+from ....ndimage import convert_fftn_to_rfftn, crop_to_shape, irfftn, make_coordinate_grid
+from ..._image_config import AbstractImageConfig
+from ..._pose import AbstractPose
+from ..base_volume import AbstractVolumeIntegrator, AbstractVoxelVolume
 
 
 try:
@@ -22,6 +24,83 @@ try:
 except ModuleNotFoundError as err:
     jnufft = None
     JAX_FINUFFT_IMPORT_ERROR = err
+
+
+class AbstractRealVoxelVolume(AbstractVoxelVolume, strict=True):
+    """Abstract interface for a voxel-based volume."""
+
+    coordinate_grid_in_pixels: eqx.AbstractVar[Float[Array, "dim dim dim 3"]]
+
+    @override
+    def rotate_to_pose(self, pose: AbstractPose, inverse: bool = False) -> Self:
+        """Return a new volume with a rotated
+        `coordinate_grid_in_pixels`.
+        """
+        return eqx.tree_at(
+            lambda d: d.coordinate_grid_in_pixels,
+            self,
+            pose.rotate_coordinates(self.coordinate_grid_in_pixels, inverse=inverse),
+        )
+
+
+class RealVoxelGridVolume(AbstractRealVoxelVolume, strict=True):
+    """A 3D voxel grid in real-space."""
+
+    real_voxel_grid: Float[Array, "dim dim dim"]
+    coordinate_grid_in_pixels: Float[Array, "dim dim dim 3"]
+
+    def __init__(
+        self,
+        real_voxel_grid: Float[NDArrayLike, "dim dim dim"],
+        coordinate_grid_in_pixels: Float[NDArrayLike, "dim dim dim 3"],
+    ):
+        """**Arguments:**
+
+        - `real_voxel_grid`: The voxel grid in fourier space.
+        - `coordinate_grid_in_pixels`: A coordinate grid.
+        """
+        self.real_voxel_grid = jnp.asarray(real_voxel_grid, dtype=float)
+        self.coordinate_grid_in_pixels = jnp.asarray(
+            coordinate_grid_in_pixels, dtype=float
+        )
+
+    @property
+    def shape(self) -> tuple[int, int, int]:
+        """The shape of the `real_voxel_grid`."""
+        return cast(tuple[int, int, int], self.real_voxel_grid.shape)
+
+    @classmethod
+    def from_real_voxel_grid(
+        cls,
+        real_voxel_grid: Float[NDArrayLike, "dim dim dim"],
+        *,
+        coordinate_grid_in_pixels: Float[Array, "dim dim dim 3"] | None = None,
+        crop_scale: float | None = None,
+    ) -> Self:
+        """Load a `RealVoxelGridVolume` from a real-valued 3D voxel grid.
+
+        **Arguments:**
+
+        - `real_voxel_grid`: A voxel grid in real space.
+        - `crop_scale`: Scale factor at which to crop `real_voxel_grid`.
+                        Must be a value greater than `1`.
+        """
+        # Cast to jax array
+        real_voxel_grid = jnp.asarray(real_voxel_grid, dtype=float)
+        # Make coordinates if not given
+        if coordinate_grid_in_pixels is None:
+            # Option for cropping template
+            if crop_scale is not None:
+                if crop_scale < 1.0:
+                    raise ValueError("`crop_scale` must be greater than 1.0")
+                cropped_shape = cast(
+                    tuple[int, int, int],
+                    tuple([int(s / crop_scale) for s in real_voxel_grid.shape[-3:]]),
+                )
+                real_voxel_grid = crop_to_shape(real_voxel_grid, cropped_shape)
+            coordinate_grid_in_pixels = make_coordinate_grid(real_voxel_grid.shape[-3:])
+
+        return cls(real_voxel_grid, coordinate_grid_in_pixels)
 
 
 class RealVoxelProjection(
