--- cryojax/simulator/_volume_integrator/__init__.py
@@ -1,12 +1,0 @@
-from .base_integrator import (
-    AbstractVolumeIntegrator as AbstractVolumeIntegrator,
-)
-from .fft_atom_projection import FFTAtomProjection as FFTAtomProjection
-from .fourier_voxel_extract import (
-    EwaldSphereExtraction as EwaldSphereExtraction,
-    FourierSliceExtraction as FourierSliceExtraction,
-)
-from .gaussian_projection import (
-    GaussianMixtureProjection as GaussianMixtureProjection,
-)
-from .real_voxel_projection import RealVoxelProjection as RealVoxelProjection
