--- cryojax/simulator/_volume_rendering/__init__.py
@@ -1,5 +1,0 @@
-from .base_rendering import AbstractVolumeRenderFn as AbstractVolumeRenderFn
-from .fft_atom_rendering import FFTAtomRenderFn as FFTAtomRenderFn
-from .gaussian_rendering import (
-    GaussianMixtureRenderFn as GaussianMixtureRenderFn,
-)
